requirements.txt
dnspython==2.6.1
Flask==3.0.3
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
